Changed project structure according to professional approach.
